+++ data/src/commonMain/kotlin/com/jotte/message/repository/WhiteboardRepository.kt
@@ -6,6 +6,6 @@ interface WhiteboardRepository {
 
     suspend fun queryWhiteboard(): WhiteboardDto
 
-    suspend fun updateWhiteboard(whiteboardDto: WhiteboardDto): Int
+    suspend fun updateWhiteboard(whiteboardDto: WhiteboardDto): Long
 
 }
